Adding data_store  and get_next function
breaking out the input and output data from file into a global variable. Creating a function to allow the next 1d input and output array to be called from the dataset - should be easy to change the get_next() for a get_live_data() or similar. The output array needs to be updated still, will use the usefulpysharkscripts.py file in L207/pysharktesting branch to complete this, ran out of time to this right now.