output data array implemented, input now reads all types
output data has the offset to error predicted time + the error type as three values.
input set to read all types not just RTPS, as large quantities of data over split packets often seems to lead into on_sample_lost errors, so these would otherwise be missed.

Thoughts on things we might need to do still:
For on_requested_deadline_missed I suspect we will need some kind of handler for matching sm_guidprefix vs its previous instance. Might need to be a handler in the input_data.py to have when the last message of that ID was received, so the AI can determine if a timeout is likely. There is surely a message response to say a timeout happened - we might be able to use this to train the system live? (possibly as part of the recursive planned nature of the AI?)